fix: update notification-service id types after MongoDB to JPA migration
NotificationService/Impl/Controller: String id -> Long id
PaymentService/Impl/Controller: Integer paymentId -> Long paymentId

Caused by entity id type change from String (MongoDB) to Long (JPA @GeneratedValue).

+++ notification-service/src/main/java/com/ecommerce/notificationservice/controller/PaymentController.java
@@ -22,7 +22,7 @@ public class PaymentController {
     }
 
     @GetMapping("/{paymentId}")
-    public Mono<Payment> getPayment(@PathVariable Integer paymentId) {
+    public Mono<Payment> getPayment(@PathVariable Long paymentId) {
         return paymentService.getPayment(paymentId);
     }
 
@@ -32,7 +32,7 @@ public class PaymentController {
     }
 
     @DeleteMapping("/{paymentId}")
-    public Mono<Void> deletePayment(@PathVariable Integer paymentId) {
+    public Mono<Void> deletePayment(@PathVariable Long paymentId) {
         return paymentService.deletePayment(paymentId);
     }
